sort_and_reindex: Make skipping optional, and document it
This fixes #45, although instead of decoupling skipping from
sort_and_reindex, this documents it and makes it a first class citizen.
This is because, after more thinking, I'm fairly certain that this is
the only place that makes sense to actually do the skipping.

There should be no behavioural changes if `skip` is not passed, since we
still check and skip by default.