be robust against millisecond strings without exactly 3 digits fixes #52
In some cases, the ms field of timestamps can have >3 digits. There
are 3 potential ways to handle such cases:

1. Throw an error. (Behavior prior to this PR)
2. Throw away the nonsignificant digits past position 3.
3. Treat the number given as the raw number of milliseconds. (This PR)

Because tools like ffmpeg and VLC go with option 3, this PR brings
srt's behavior to parity with what is arguably the "gold standard".

In general, these gold standard tools only look at the raw number
in each field and ignore any positional information. This PR also
introduces tests to ensure this behavior is preserved by srt.

I.e., a timestamp like

00:115:05,12

is the same as

01:55:05,012

and not

00:11:35,120

That is, we parse the minutes field as "115 minutes" and not "11.5
minutes", and we parse the milliseconds field as "12 milliseconds" and
not "120 milliseconds" as if the rightmost zero were accidentally
truncated. In summary, it's the value at each field that matters, not
the position of individual digits with each field.