test_srt: Fix rare flakiness in test_parser_noncontiguous_leading
During release testing, Hypothesis running in Travis
(https://travis-ci.org/cdown/srt/jobs/622124156) reported the following
failing case on OSX:

    @example(
        subs=[
            srt.Subtitle(
                index=0,
                start=timedelta(0),
                end=timedelta(21186, 33261),
                content="0",
                proprietary="",
            )
        ],
        garbage=" -",
    )

This rarely comes up since Hypothesis usually won't have time to
generate a failing example of this kind. The fix is to just check if the
last character is a dash, since if it's earlier, it will genuinely be
invalid since it can't be parsed as part of a digit, which is checked
for in the following check.